# src/swagger.yaml
swagger: '2.0'
info:
  description: Tài liệu API cho Hệ thống quản lý học viên
  version: '1.0.0'
  title: Hệ thống quản ý Học viên
  contact:
    email: trinhthang1704@gmail.com
schemes:
  - http
  - https
security:
  - Bearer: []
basePath: /
paths:
...
  /schedule:
    post:
      tags:
        - schedule
      summary: Tạo thời khóa biểu 
      x-swagger-router-controller: Schedule
      operationId: createSchedule
      parameters:
      - $ref: '#/parameters/CreateScheduleParameter'
      produces:
        - application/json
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
    
    get:
      tags:
        - schedule
      summary: Lấy danh sách thời khóa biểu
      x-swagger-router-controller: Schedule
      operationId: getSchedules
      parameters:
      - $ref: '#/parameters/startDateParameter'
      - $ref: '#/parameters/endDateParameter'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/SchedulesCollection'
        '400':
          description: Bad Request
  /attendence/upload:
    post:
      tags:
        - attendence
      summary: Tạo dữ liệu điểm danh 
      x-swagger-router-controller: Attendence
      operationId: createAttendence
      parameters:
      - $ref: '#/parameters/CreateStudentsParameter'
      produces:
        - application/json
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
  /attendence:
    get:
      tags:
        - attendence
      summary: Lấy danh sách dữ liệu điêm danh 
      x-swagger-router-controller: Attendence
      operationId: getAttendences
      parameters:
      - $ref: '#/parameters/SearchNameParameter'
      - $ref: '#/parameters/ClassIdsParameter'
      - $ref: '#/parameters/DateParameter'
      - $ref: '#/parameters/LimitParameter'
      - $ref: '#/parameters/OffsetParameter'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/SesionsCollection'
        '400':
          description: Bad Request
  /attendence/attendenceStats:
    get:
      tags:
        - attendence
      summary: Lấy chỉ số và dữ liệu biểu đồ điểm danh
      x-swagger-router-controller: Attendence
      operationId: getAttendenceStats
      parameters:
      - $ref: '#/parameters/SchoolYearIdParameter'
      - $ref: '#/parameters/StartDateParameter'
      - $ref: '#/parameters/EndDateParameter'
      - $ref: '#/parameters/ClassIdParameter'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/StatCollection'
        '400':
          description: Bad Request
  /attendence/topAbsent:
    get:
      tags:
        - attendence
      summary: Lấy top 10 học viên nghỉ học nhiều nhất
      x-swagger-router-controller: Attendence
      operationId: getTopAbsent
      parameters:
      - $ref: '#/parameters/SchoolYearIdParameter'
      - $ref: '#/parameters/ClassIdParameter'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/SesionsCollection'
        '400':
          description: Bad Request
...
parameters:
  SchoolYearIdParameter:
    in: query
    name: schoolYearId
    type: number
    required: true
  StartDateParameter:
    in: query
    name: startDate
    type: string
  EndDateParameter:
    in: query
    name: endDate
    type: string
  ClassIdParameter:
    in: query
    name: classId
    type: number
  SearchNameParameter:
    in: query 
    name: searchName
    type: string
  ClassIdsParameter:
    in: query
    name: classIds
    type: array
    items:
      type: string
  DateParameter:
    in: query
    name: date
    type: string
  LimitParameter:
    in: query
    name: limit
    type: number
    default: 10
  OffsetParameter:
    in: query
    name: offset
    type: number
    default: 0
  startDateParameter:
    in: query
    name: startDate
    type: string
  endDateParameter:
    in: query
    name: endDate
    type: string
  CreateScheduleParameter:
    in: body
    name: body
    required: true
    schema:
      $ref: "#/definitions/CreateScheduleRequest"
...
  CreateStudentsParameter:
    in: formData
    name: file
    type: file
    description: file được tải lên.
...
definitions:
  StatCollection:
    type: object
    properties:
      stat: 
        type: object
        properties:
          total: 
            type: number
          attend: 
            type: object
            properties:
              value:
                type: number
              percent:
                type: number
          absent: 
            type: object
            properties:
              value:
                type: number
              percent:
                type: number
          late: 
            type: object
            properties:
              value:
                type: number
              percent:
                type: number
      charts: 
        type: array
        items:
          type: object
          properties:
            id: 
              type: number
            name: 
              type: string
            value:
              type: number
  SchedulesCollection:
    type: array
    items:
      type: object
      properties:
        subject:
          type: string
        class:
          type: string
        classroom:
          type: string
        date:
          type: string
        session:
          type: object
          properties:
            title:
              type: string
            startTime:
              type: string
            endTime:
              type: string
        category:
          type: string
        lession:
          type: string
        teacher:
          type: object
          properties:
            name:
              type: string
            phone:
              type: string
  CreateScheduleRequest:
    type: object 
    properties:
      subjectId:
        type: number
      categoryId:
        type: number
      classId:
        type: number
      learningDate:
        type: string
      sessionId:
        type: number
      accountId:
        type: number
      classroomId:
        type: number
      startDate:
        type: string
      endDate:
        type: string
      finalExamDate:
        type: string
  SesionsCollection:
    type: array
    items:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        startTime:
          type: string
        endTime:
          type: string
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: 'Example: "Bearer ABC123"'